edit strlen protype

+++ ft_strlen.c
@@ -12,7 +12,7 @@
 
 #include "libft.h"
 
-int	ft_strlen(char *str)
+size_t	ft_strlen(char *str)
 {
 	int count;
